Add files via upload
It's all about including a bunch of .h files and not knowing where funcX(blabla) comes from. It also clearly identifies funcX if in different files a funcX exists. It is what we call 'namespacing' or 'qualifying' or 'topologying' an object. And it is my deepest wish that libc eventually will be implemented this way.

So, someone calling a func coming from easystring.c will be forced to call that function by it's external funcpointer str.funcX
example :  
       string s = str.new("Hello world."); 
       str.repl( s, "world","everybody");

I don't think this disadvantages execution. It only asks for an initialisation at startpoint of the program : strClass_init();